requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
antlr4-python3-runtime==4.9.3
anyio==3.7.0
appdirs==1.4.4
arrow==1.2.3
async-timeout==4.0.2
asyncssh==2.13.1
atpublic==3.1.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==3.6.4.0
blessed==1.20.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.26.3
colorama==0.4.6
configobj==5.0.8
croniter==1.3.15
cryptography==41.0.1
dateutils==0.6.12
deepdiff==6.3.0
dictdiffer==0.9.0
diskcache==5.6.1
distro==1.8.0
dpath==2.1.6
dulwich==0.21.5
dvc==2.58.2
dvc-data==0.51.0
dvc-http==2.30.2
dvc-objects==0.22.0
dvc-render==0.5.3
dvc-studio-client==0.10.0
dvc-task==0.2.1
exceptiongroup==1.1.1
fastapi==0.88.0
filelock==3.12.0
flatten-dict==0.4.2
flufl.lock==7.1.1
frozenlist==1.3.3
fsspec==2023.5.0
funcy==2.0
gitdb==4.0.10
GitPython==3.1.31
grandalf==0.8
h11==0.14.0
hydra-core==1.3.2
idna==3.4
inquirer==3.1.3
iterative-telemetry==0.0.8
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.0
kornia==0.6.12
lightning==2.0.2
lightning-cloud==0.5.36
lightning-utilities==0.8.0
lit==16.0.5.post0
markdown-it-py==2.2.0
MarkupSafe==2.1.3
mdurl==0.1.2
monai==1.1.0
mpmath==1.3.0
multidict==6.0.4
nanotime==0.5.2
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
ordered-set==4.1.0
orjson==3.9.0
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
platformdirs==3.5.1
prompt-toolkit==3.0.38
psutil==5.9.5
pycparser==2.21
pydantic==1.10.8
pydot==1.4.2
pygit2==1.12.1
Pygments==2.15.1
pygtrie==2.5.0
PyJWT==2.7.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.2
pytz==2023.3
PyYAML==6.0
readchar==4.0.5
requests==2.31.0
rich==13.4.1
ruamel.yaml==0.17.31
ruamel.yaml.clib==0.2.7
scmrepo==1.0.3
shortuuid==1.0.11
shtab==1.6.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
sqltrie==0.3.1
starlette==0.22.0
starsessions==1.3.0
sympy==1.12
tabulate==0.9.0
tomlkit==0.11.8
torch==2.0.1
torchmetrics==0.11.4
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
uvicorn==0.22.0
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
websocket-client==1.5.2
websockets==11.0.3
yarl==1.9.2
zc.lockfile==3.0.post1
